feat

+++ internal/entity/user.go
@@ -4,5 +4,5 @@ type User struct {
 	ID       int     `json:"id"`
 	Login    string  `json:"login"`
 	Password string  `json:"password"`
-	Balance  float64 `json:"balance"`
+	Balance  float32 `json:"balance"`
 }
